+++ src/Log_node.cc
@@ -35,7 +35,7 @@ void Log_node::start() {
   }
 }
 
-void Log_node::hook_added_data_reader(int reader) {
+void Log_node::hook_added_data_reader(size_t reader) {
 }
-void Log_node::hook_added_data_writer(int writer) {
+void Log_node::hook_added_data_writer(size_t writer) {
 }
